+++ listar/veiculos.php
@@ -1,17 +1,6 @@
 <?php
 require "configs/functions.php";
 
-// Verificar se a categoria do usuário é "funcionário" para permitir o acesso
-if ($_SESSION["usuarioAdm"]["categoria"] !== "Funcionario") {
-    // Caso não seja Funcionario
-    mensagem("Erro", "Voce não tem permição para acessar essa pagina");
-
-    exit;
-}
-// Caso seja funcionario
-
-if (!isset($pagina))
-    exit;
 ?>
 
 <div class="cadastramento">
@@ -34,7 +23,7 @@ if (!isset($pagina))
                     <tr>
                         <td>Modelo do Veiculo</td>
                         <td>Placa</td>
-                        <td width="100px">Excluir</td>
+                        <td width="100px">Editar/Excluir</td>
 
                     </tr>
                 </thead>
@@ -50,12 +39,17 @@ if (!isset($pagina))
                             <td><?= $d->Modelo ?></td>
                             <td><?= $d->placa ?></td>
                             <td class="text-center">
-                                <a href="cadastrar/veiculos/<?= $d->id ?>" title="Editar" class="btn btn-success btn-sm">
-                                    <i class="fas fa-edit"></i>
-                                </a>
-                                <a href="javascript:excluir(<?= $d->id ?>)" title="Excluir" class="btn btn-danger btn-sm">
-                                    <i class="fas fa-trash"></i>
-                                </a>
+                                <?php if ($_SESSION["usuarioAdm"]["categoria"] === "Funcionario") { ?>
+                                    <a href="cadastrar/veiculos/<?= $d->id ?>" title="Editar" class="btn btn-success btn-sm">
+                                        <i class="fas fa-edit"></i>
+                                    </a>
+                                    <a href="javascript:excluir(<?= $d->id ?>)" title="Excluir" class="btn btn-danger btn-sm">
+                                        <i class="fas fa-trash"></i>
+                                    </a>
+                                <?php } else { ?>
+                                    <!-- Apenas exibição, sem opções de editar/excluir -->
+                                    <span class="text-muted">Não Habilitado</span>
+                                <?php } ?>
                             </td>
                         </tr>
